Fix files to be PSR-4 compatible
Previously, If you try to load NylasModelCollection before Nylas, autoloader fails
